% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{as.eyer.pupilr}
\alias{as.eyer.pupilr}
\title{Converts pupilr object and all nested object to an eyer objects}
\usage{
\method{as.eyer}{pupilr}(obj, ...)
}
\arguments{
\item{obj}{surface.item object}
}
\description{
Converts pupilr object and all nested object to an eyer objects
}
