% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{as.eyer}
\alias{as.eyer}
\title{Converts given object to an eyer object}
\usage{
as.eyer(obj, ...)
}
\arguments{
\item{obj}{Object to be converted}

\item{...}{}
}
\value{
valid eyer object, list of eyer objects or NULL if it can't be converted
}
\description{

}
