% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{as.eyer.surface.item}
\alias{as.eyer.surface.item}
\title{Converts surface item to an eyer object}
\usage{
\method{as.eyer}{surface.item}(obj, ...)
}
\arguments{
\item{obj}{surface.item object}

\item{...}{}
}
\value{
EyerObject
}
\description{
Converts surface item to an eyer object
}
