% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{as.eyer.surfaces}
\alias{as.eyer.surfaces}
\title{Converts all nested surfaces objects to eyer objects}
\usage{
\method{as.eyer}{surfaces}(obj, ...)
}
\arguments{
\item{obj}{surface.item object}
}
\description{
Converts all nested surfaces objects to eyer objects
}
