% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{calculate_adjustedRT_f}
\alias{calculate_adjustedRT_f}
\title{Calculated adjusted RTs}
\usage{
calculate_adjustedRT_f(df)
}
\arguments{
\item{df}{a data frame (pre-processed by \code{pupilr})}
}
\value{
the input data frame, with the new columns, \code{meanBaselineRT}, \code{AdjustedRT}, and \code{meanAdjustedRT}, the last of which represents the participant's mean adjusted RT for the experimental trials
}
\description{
Calculates each subject's mean Baseline RT and subtracts that value from
the experimental RTs.
}

