% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{change_timestamps_start.pupilr}
\alias{change_timestamps_start.pupilr}
\title{Changes timestamps to a given start time}
\usage{
\method{change_timestamps_start}{pupilr}(obj, start_time)
}
\description{
Changes timestamps to a given start time
}
