% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{change_timestamps_start.surface.item}
\alias{change_timestamps_start.surface.item}
\title{Changes timestamps to a given start time}
\usage{
\method{change_timestamps_start}{surface.item}(obj, start_time, new_start_time = start_time)
}
\description{
Changes timestamps to a given start time
}
