% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{change_timestamps_start.surface}
\alias{change_timestamps_start.surface}
\title{So far changes ONLY GAZE property}
\usage{
\method{change_timestamps_start}{surface}(obj, start_time)
}
\arguments{
\item{start_time}{}
}
\description{
So far changes ONLY GAZE property
}
