% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{convert_to_eyer}
\alias{convert_to_eyer}
\title{Converts given object to an eyer object}
\usage{
convert_to_eyer(obj, ...)
}
\arguments{
\item{...}{}
}
\description{
Converts given object to an eyer object
}
