% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_level_exclusions.R
\name{exclude_extreme_rts_f}
\alias{exclude_extreme_rts_f}
\title{Remove trial-level exclusions}
\usage{
exclude_extreme_rts_f(df)
}
\arguments{
\item{df}{the data frame (pre-processed by \code{pupilr})}
}
\value{
the same data frame, but without the excluded RT
}
\description{
This excludes trials with RTs < 200 ms and RTs with that are greater than 3 SDs
from the worker's mean RT.
}

