% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{filter_times.surface}
\alias{filter_times.surface}
\title{Filters out only times between start and end}
\usage{
\method{filter_times}{surface}(obj, start, end, ...)
}
\arguments{
\item{obj}{surface object}

\item{start}{start time in seconds}

\item{end}{end time in seconds}

\item{...}{optional parameters just like in get_gaze_timewindow, get_fixations (e.g. on_surface)}
}
\value{
surface object with filtered out times
}
\description{
Filters out only times between start and end
}
