% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_artifacts.R
\name{find_artifacts}
\alias{find_artifacts}
\title{Check for outlier created through artifacts}
\usage{
find_artifacts(pupil_size, tol = 3)
}
\arguments{
\item{pupil_size}{input: numeric vector of pupil sizes}

\item{tol}{input: numeric value for spread tolerance. Actually it is the factor the standard deviation is multiplied by; Default=3}
}
\value{
logical vector indicating, whether data point is an outlier. Value is TRUE, when
pupil size sample is bigger than the mean of the whole vector + \code{tol} times the sd of the whole vector
pupil size sample is smaller than the mean of the whole vector - \code{tol} times the sd of the whole vector
}
\description{
\code{find_artifact} checks an numeric vector for outliers, which deviate from the mean by more than a variable multiple of the standard deviation.
}
\examples{
x <- rnorm(100, 5, 3)
x[10] <- 80
find_artifacts(x)

}

