% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-getters.R
\name{get_gaze.surface}
\alias{get_gaze.surface}
\title{Returns gaze data.frame from the surface object}
\usage{
\method{get_gaze}{surface}(obj, on_surface = T, ...)
}
\arguments{
\item{obj}{surface object}

\item{on_surface}{should it only return those measurements on surface?}

\item{...}{}
}
\value{
data.frame with gaze
}
\description{
Returns gaze data.frame from the surface object
}
