% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-getters.R
\name{get_gaze_timewindow.surface}
\alias{get_gaze_timewindow.surface}
\title{returns gaze data.frame between given times}
\usage{
\method{get_gaze_timewindow}{surface}(obj, start, end, since_start = F,
  ...)
}
\arguments{
\item{obj}{surface object}

\item{start}{start time in seconds (numeric)}

\item{end}{end time in seconds (numeric)}

\item{since_start}{defaults to false}

\item{...}{optional parameters, same as in get_gaze}
}
\value{
filtered data.frame
}
\description{
returns gaze data.frame between given times
}
\examples{
gaze <- get_gaze_timewindow(surface_object, 1000, 1200, on_surface = F)
}
