% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pupil_average.R
\name{get_pupil_avg}
\alias{get_pupil_avg}
\title{Title Average a pupil size vector around a peak pupil size}
\usage{
get_pupil_avg(pupil_size, avg_window = c(-48L, 48L), region)
}
\arguments{
\item{pupil_size}{input: numeric vector of pupil sizes}

\item{avg_window}{input: Two element vector of integers determining the number of samples before and after the peak pupil size, over which the average is calculated; Default value adds up to around 200ms before and after peak pupil size, when sampling with 250 Hz.}

\item{region}{Two element vector containing the range of samples to use; keeps the function flexible, so that it can be used for baseline and recall measure.}
}
\value{
numeric object: Average pupil size round peak pupil size.
}
\description{
\code{get_pupil_avg} gets the mean of pupil size samples around the peak pupil size. Actually this function is used to create the measure for inferential statistic analysis
}
\examples{
x <- rnorm(100, 5, 3)
x[50] <- 80
get_pupil_avg(x, c(-5,5), 1:100)
}

