% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-getters.R
\name{get_surface}
\alias{get_surface}
\title{Returns sufface of a given name}
\usage{
get_surface(obj, surface_name)
}
\arguments{
\item{obj}{pupilr object}

\item{surface_name}{name of the surface to be extracted}
}
\value{

}
\description{
Returns sufface of a given name
}
