% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{initialize_scores}
\alias{initialize_scores}
\title{Initialize scores for data}
\usage{
initialize_scores(df)
}
\arguments{
\item{df}{a data frame (pre-processed by \code{pupilr})}
}
\value{
the input data frame but with the new columns
}
\description{
Adds/overwrites columns: \code{Phase}, \code{subjMean_BlockRT}, \code{subjMean_BlockAcc},
\code{BlockmeanRT}, and \code{Rt.Z}
}

