% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-getters.R
\name{is.pupilr}
\alias{is.pupilr}
\title{Returns if the object has class pupilr}
\usage{
is.pupilr(obj)
}
\arguments{
\item{obj}{object ot be evaluated}
}
\value{

}
\description{
Returns if the object has class pupilr
}
