% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_and_overwrite.R
\name{left_join_and_overwrite}
\alias{left_join_and_overwrite}
\title{Left join and overwrite values in "left" dataframe}
\usage{
left_join_and_overwrite(x, y, by = NULL, copy = FALSE,
  warnifnoprevious = FALSE, ...)
}
\arguments{
\item{x, y}{the "left" and "right" data frames, respectively}

\item{by}{the columns you want to join by would appear by default.}

\item{copy}{whatever this argument means in the regular \code{\link[dplyr]{left_join}}}

\item{\dots}{Optional arguments}
}
\value{
a data frame or tibble or whatever, just like \code{\link[dplyr]{left_join}} but with columns overwritten
}
\description{
If you want to recalculate Z-scores, for example, without worrying about columns in the
"right" dataframe having the same names as columns in the left data frame
(which would cause their names to become something like "colname.x" and "colname.y")
this function replaces all the columns in the first data frame with the ones in the second.
The arguments are just like \code{\link[dplyr]{left_join}}
}

