% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{load_key_value_file}
\alias{load_key_value_file}
\title{Function to read pupil key, value style dataframes into named list}
\usage{
load_key_value_file(filepath)
}
\arguments{
\item{filepath}{path to the file}
}
\value{
loaded named list
}
\description{
Loads the dataframe and returns named list
}
