% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{load_surface_data}
\alias{load_surface_data}
\title{Loads folder with surface data}
\usage{
load_surface_data(dir)
}
\arguments{
\item{dir}{path to the folder}
}
\description{
Loads folder with surface data
}
