% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_level_exclusions.R
\name{non_eligible_cheaters_f}
\alias{non_eligible_cheaters_f}
\title{Identify participant-level exclusions based on performance factors: 'cheating'}
\usage{
non_eligible_cheaters_f(df)
}
\arguments{
\item{df}{a data frame (pre-processed by \code{pupilr})}
}
\value{
a data frame of workers who need to be excluded
}
\description{
Corresponds to Step 1.2.1:
\cr
identify subjects who have mean block RTs < 200ms
}

