% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_level_exclusions.R
\name{non_eligible_participants_f}
\alias{non_eligible_participants_f}
\title{Identify participants to be excluded}
\usage{
non_eligible_participants_f(df, duplicate_turkers, newsurvey = TRUE)
}
\arguments{
\item{duplicate_turkers}{a data frame of duplicate participants}

\item{newsurvey}{a boolean that is \code{TRUE} when the data frame has the updated survey questions}

\item{d}{a data frame}
}
\value{
a data frame of workers who need to be excluded, with reasons for exclusion
}
\description{
Exclude all participants with the Step 1 ("participant-level") exclusion criteria:
\describe{
  \item{(i)}{not monolingual English speakers}
  \item{(ii)}{did not use appropriate audio equipment for the study}
  \item{(iii)}{reported high familiarity with accent}
  \item{(iv)}{subjects whose mean RT in any block was > 3 SDs from Condition mean}
  \item{(v)}{subjects who had a mean RT in any block < 200 ms}
}
}
\details{
\strong{NOTE:} We did \emph{not} exclude participants based on mean RTs in the practice block, whereas our previous work had.
Additionally, our previous work \emph{did not exclude} participants with block means < 200ms
\cr
\strong{NOTE:} We did \emph{NOT} exclude participants for reasons 1-4 before calculating the standard deviation for reason 5
}

