% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_level_exclusions.R
\name{non_eligible_step1_f}
\alias{non_eligible_step1_f}
\title{Identify participant-level exclusions based on participant factors}
\usage{
non_eligible_step1_f(d, duplicate_turkers, newsurvey = TRUE)
}
\arguments{
\item{d}{a data frame (pre-processed by \code{pupilr})}

\item{duplicate_turkers}{a data frame of duplicate participants}

\item{newsurvey}{a boolean that is \code{TRUE} when the data frame has the updated survey questions}
}
\value{
a data frame of workers who need to be excluded
}
\description{
Corresponds to Step 1.1:
\cr
identify subjects who do not meet participation criteria:
\describe{
  \item{(i)}{not monolingual English speakers}
  \item{(ii)}{did not use appropriate audio equipment for the study}
  \item{(iii)}{reported high familiarity with accent }
}
}

