% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/participant_level_exclusions.R
\name{non_eligible_step2_f}
\alias{non_eligible_step2_f}
\title{Identify participant-level exclusions based on performance factors: outliers}
\usage{
non_eligible_step2_f(df)
}
\arguments{
\item{df}{a data frame (pre-processed by \code{pupilr})}
}
\value{
a data frame of workers who need to be excluded
}
\description{
Corresponds to Step 1.2.2:
\cr
identify *eligible* subjects who had a mean block RT > 3 SDs from their Condition block mean.
}

