% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_fixations_file}
\alias{open_fixations_file}
\title{Retusns fixations data.frame}
\usage{
open_fixations_file(folder)
}
\arguments{
\item{folder}{directory with the exported data}
}
\value{
data.frame with loaded fixations positions
}
\description{
returns loaded fixations file from the folder,
in case you exported it as part of the export
}
