% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_gaze_file}
\alias{open_gaze_file}
\title{Retusns gaze positions data frame}
\usage{
open_gaze_file(folder)
}
\arguments{
\item{folder}{directory with the exported data}
}
\value{
data.frame with loaded gaze positions
}
\description{
Retusns gaze positions data frame
}
