% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_info_file}
\alias{open_info_file}
\title{Finds and loads \code{info.csv} file from given directory}
\usage{
open_info_file(folder)
}
\arguments{
\item{folder}{directory with the `info.csv`` file}
}
\value{
list with info file contents
}
\description{
Finds and loads \code{info.csv} file from given directory
}
