% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_positions_file}
\alias{open_positions_file}
\title{Returns data frame of pupil positions}
\usage{
open_positions_file(dir)
}
\arguments{
\item{dir}{directory with the exported data}
}
\value{
data.frame with loaded pupil positions
}
\description{
Returns data frame of pupil positions
}
