% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_surfaces}
\alias{open_surfaces}
\title{Searches for the surfaces and tries to load them into a list}
\usage{
open_surfaces(folder)
}
\arguments{
\item{folder}{directory with the exported data}
}
\value{
list with surfaces and their definitions
}
\description{
Searches for the surfaces and tries to load them into a list
}
