% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_timestamps_file}
\alias{open_timestamps_file}
\title{Returns data frame of world_timestamps}
\usage{
open_timestamps_file(folder)
}
\arguments{
\item{folder}{directory with the exported data}
}
\value{
data.frame with loaded timestamps
}
\description{
Returns data frame of world_timestamps
}
