% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-visualising.R
\name{plot_eye_heatmap}
\alias{plot_eye_heatmap}
\title{Generic function to plot eye heatmap given passed x and y coordinates}
\usage{
plot_eye_heatmap(x, y)
}
\arguments{
\item{x}{x coordinates}

\item{y}{y coordinates}
}
\value{
ggplot og stat density
}
\description{
Generic function to plot eye heatmap given passed x and y coordinates
}
