% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-visualisations.R
\name{plot_gaze_data}
\alias{plot_gaze_data}
\title{Pot gaze data}
\usage{
plot_gaze_data(obj, downsample_rate = 0, surface_name = NULL)
}
\arguments{
\item{obj}{pupilr object}

\item{downsample_rate}{rate of downsample (2 = 50 percent, 10 = 10 percent)}

\item{surface_name}{}
}
\value{

}
\description{
Pot gaze data
}
