% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-visualising.R
\name{plot_gaze_heatmap}
\alias{plot_gaze_heatmap}
\title{Plots heatmap of the gaze data}
\usage{
plot_gaze_heatmap(obj)
}
\arguments{
\item{obj}{either pupilr object or surface object}
}
\description{
Plots heatmap of the gaze data
}
