% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-visualising.R
\name{plot_gaze_heatmap.surface}
\alias{plot_gaze_heatmap.surface}
\title{PLots heatmap of a surface object}
\usage{
\method{plot_gaze_heatmap}{surface}(obj)
}
\arguments{
\item{obj}{object of type surface}
}
\value{
ggplot of stat density type with given limits
}
\description{
PLots heatmap of a surface object
}
