% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_transforms.R
\name{redo_Z_scores}
\alias{redo_Z_scores}
\title{Calculate Z-scores for mean block RTs}
\usage{
redo_Z_scores(df)
}
\arguments{
\item{df}{a data frame (pre-processed by \code{pupilr})}
}
\value{
the input data frame, with the new columns, \code{BlockmeanRT}, \code{Rt.Z}
}
\description{
Calculates the z-scores \strong{within Conditions} and within block number for participants' mean block RT.
Will overwrite any columns named \code{BlockmeanRT} or \code{Rt.Z}.
}

