% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{remove_not_on_surface}
\alias{remove_not_on_surface}
\title{Removes measurements which are not on given surface}
\usage{
remove_not_on_surface(obj)
}
\value{
modified object
}
\description{
Removes measurements which are not on given surface
}
