% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{remove_not_on_surface.surface}
\alias{remove_not_on_surface.surface}
\title{Removes measurements which are not on given surface}
\usage{
\method{remove_not_on_surface}{surface}(obj)
}
\arguments{
\item{obj}{surface object}
}
\value{
modified object
}
\description{
Removes measurements which are not on given surface
}
