% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check_data.R
\name{sanity_check_data}
\alias{sanity_check_data}
\title{Does a few sanity checks on the experimental data}
\usage{
sanity_check_data(df, pilot = TRUE)
}
\arguments{
\item{df}{The data frame}
}
\value{
Nothing
}
\description{
Mostly designed as a sanity check for the pilot data so far
}

