% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-object.R
\name{SurfaceItemObject}
\alias{SurfaceItemObject}
\title{Creates surface.item object}
\usage{
SurfaceItemObject()
}
\value{
object of class \code{surface.item}
}
\description{
This object contains basic eye information about the particular surface.

SurfaceItemObject:
\itemize{
\item data: gaze, fixations and events for given surface
}
}
