% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-object.R
\name{SurfacesObject}
\alias{SurfacesObject}
\title{Creates surfaces object}
\usage{
SurfacesObject()
}
\value{
object of class \code{surfaces}
}
\description{
Surfaces object contains basic data from the surfaces folder
(generally events table with information about surfaces enter adn exit).

SurfaceObject:
\itemize{
\item data: data about all surfaces together
\item items: list with named field for each surface containig `SurfaceItemObject``
}
}
