% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-preprocessing.R
\name{filter_times}
\alias{filter_times}
\title{Filters out object with only times between start and end}
\usage{
filter_times(obj, start, end, ...)
}
\arguments{
\item{obj}{object to be modified}

\item{start}{start time in seconds}

\item{end}{end time in seconds}

\item{...}{}
}
\description{
Filters out object with only times between start and end
}
