% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-getters.R
\name{get_surface_names}
\alias{get_surface_names}
\title{Returns names of all surfaces in the object}
\usage{
get_surface_names(obj)
}
\arguments{
\item{obj}{pupilr object}
}
\value{
character(0) if no surfaces are found or vector of character names
}
\description{
Returns names of all surfaces in the object
}
