% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{load_folder}
\alias{load_folder}
\title{Loads all important exported objecs from given folder}
\usage{
load_folder(dir)
}
\arguments{
\item{dir}{directory to load}
}
\value{
PupilrObject
}
\description{
Loads all important exported objecs from given folder
}
