% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{load_surface_data}
\alias{load_surface_data}
\title{Loads folder with surface data}
\usage{
load_surface_data(folder, preprocess = T)
}
\arguments{
\item{folder}{path to the folder}

\item{preprocess:}{if the data should be immediately cleaned. default T. If set to false, some filtering and plotting functions will not work}
}
\value{

}
\description{
Loads folder with surface data
}
