% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_exported_file}
\alias{open_exported_file}
\title{General function to load preprocessed file based on pattern}
\usage{
open_exported_file(dir, ptr)
}
\arguments{
\item{dir}{where to look for the data.frame csv}

\item{ptr}{patten to search for}
}
\value{
data.frame
}
\description{
General function to load preprocessed file based on pattern
}
