% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pupilr-loading.R
\name{open_info_file}
\alias{open_info_file}
\title{Finds and loads info file from given directory}
\usage{
open_info_file(dir)
}
\arguments{
\item{dir}{directory in which to search}
}
\value{
loaded info data.frame
}
\description{
Finds and loads info file from given directory
}
